pull out shared socrata functionality
